# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Café Lauren",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/slots/{slot_id}/keep": {
      "post": {
        "tags": [
          "slots"
        ],
        "summary": "Keep",
        "operationId": "keepSlot",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Week"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/slots/{slot_id}/vote": {
      "post": {
        "tags": [
          "slots"
        ],
        "summary": "Vote",
        "operationId": "voteSlot",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Week"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/slots/{slot_id}/swap": {
      "post": {
        "tags": [
          "slots"
        ],
        "summary": "Swap",
        "operationId": "swapSlot",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Week"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/slots/{slot_id}/swap-options": {
      "post": {
        "tags": [
          "slots"
        ],
        "summary": "Swap Options",
        "operationId": "swapOptions",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapOptionsRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/slots/{slot_id}/reject": {
      "post": {
        "tags": [
          "slots"
        ],
        "summary": "Reject",
        "operationId": "rejectSlot",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/slots/{slot_id}/move": {
      "post": {
        "tags": [
          "slots"
        ],
        "summary": "Move",
        "operationId": "moveSlot",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Week"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/slots/{slot_id}/cook": {
      "post": {
        "tags": [
          "slots"
        ],
        "summary": "Set Cook",
        "operationId": "setSlotCook",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Week"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/slots/{slot_id}": {
      "patch": {
        "tags": [
          "slots"
        ],
        "summary": "Patch Slot",
        "operationId": "patchSlot",
        "parameters": [
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slot Id"
            }
          },
          {
            "name": "x-cafe-user",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "lauren | joe | leidy",
              "title": "X-Cafe-User"
            },
            "description": "lauren | joe | leidy"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Week"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CookRequest": {
        "properties": {
          "cook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cook",
            "description": "Who cooks: a person key or free text."
          }
        },
        "type": "object",
        "title": "CookRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Ingredient": {
        "properties": {
          "qty": {
            "type": "string",
            "title": "Qty",
            "default": ""
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Ingredient"
      },
...
      "JobAccepted": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "type": "object",
        "required": [
          "job"
        ],
        "title": "JobAccepted"
      },
...
      "MoveRequest": {
        "properties": {
          "to": {
            "type": "string",
            "enum": [
              "mon",
              "tue",
              "wed",
              "thu",
              "fri",
              "sat",
              "sun"
            ],
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "to"
        ],
        "title": "MoveRequest"
      },
...
      "RejectRequest": {
        "properties": {
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "remember": {
            "type": "boolean",
            "title": "Remember",
            "default": false
          },
          "mode": {
            "type": "string",
            "enum": [
              "open",
              "another"
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "RejectRequest"
      },
      "RejectResponse": {
        "properties": {
          "week": {
            "$ref": "#/components/schemas/Week"
          },
          "feedback_id": {
            "type": "integer",
            "title": "Feedback Id"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Job"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "week",
          "feedback_id"
        ],
        "title": "RejectResponse"
      },
...
      "Slot": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "day": {
            "type": "string",
            "enum": [
              "mon",
              "tue",
              "wed",
              "thu",
              "fri",
              "sat",
              "sun"
            ],
            "title": "Day"
          },
          "kind": {
            "type": "string",
            "enum": [
              "cook",
              "leftover",
              "leidy",
              "custom",
              "open"
            ],
            "title": "Kind"
          },
          "recipe_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Id"
          },
          "recipe": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "suggested",
                  "thinking",
                  "kept",
                  "edited",
                  "approved",
                  "rejected"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "By"
          },
          "basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basis"
          },
          "why": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Why"
          },
          "cook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cook"
          },
          "ingredient_flags": {
            "additionalProperties": {
              "$ref": "#/components/schemas/IngredientFlag"
            },
            "type": "object",
            "title": "Ingredient Flags"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/SlotIngredient"
            },
            "type": "array",
            "title": "Ingredients",
            "description": "Effective ingredients (week override or recipe) with tags."
          },
          "ingredients_edited": {
            "type": "boolean",
            "title": "Ingredients Edited",
            "default": false
          },
          "votes": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "up",
                "down"
              ]
            },
            "type": "object",
            "title": "Votes",
            "description": "person key -> up/down"
          }
        },
        "type": "object",
        "required": [
          "id",
          "day",
          "kind"
        ],
        "title": "Slot"
      },
...
      "SlotPatch": {
        "properties": {
          "ingredients": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Ingredient"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients"
          },
          "ingredient_flags": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/IngredientFlag"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredient Flags"
          },
          "reset_ingredients": {
            "type": "boolean",
            "title": "Reset Ingredients",
            "default": false
          }
        },
        "type": "object",
        "title": "SlotPatch",
        "description": "Meal-detail edits for this week only. Marks a cook slot as edited."
      },
...
      "Store": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ads_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ads Url"
          },
          "instacart_retailer_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instacart Retailer Key"
          },
          "deal_count": {
            "type": "integer",
            "title": "Deal Count",
            "default": 0
          },
          "ad_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad From"
          },
          "ad_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad To"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "name"
        ],
        "title": "Store"
      },
...
      "SwapOptionsRequest": {
        "properties": {
          "prefs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prefs"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "type": "object",
        "title": "SwapOptionsRequest"
      },
      "SwapRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "recipe",
              "text",
              "leftover",
              "leidy",
              "open"
            ],
            "title": "Kind"
          },
          "recipe_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Id"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basis",
            "description": "What the person asked for, echoed in the UI."
          },
          "cook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cook"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "SwapRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoteRequest": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "up",
                  "down"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "null clears your vote"
          }
        },
        "type": "object",
        "title": "VoteRequest"
      },
      "Week": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "monday": {
            "type": "string",
            "format": "date",
            "title": "Monday"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "\"Week of August 24\""
          },
          "store": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Store"
              },
              {
                "type": "null"
              }
            ]
          },
          "order_via": {
            "type": "string",
            "enum": [
              "delivery",
              "pickup",
              "amazon",
              "share",
              "self"
            ],
            "title": "Order Via"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/Slot"
            },
            "type": "array",
            "title": "Slots"
          },
          "list_diff_count": {
            "type": "integer",
            "title": "List Diff Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "monday",
          "label",
          "order_via",
          "slots"
        ],
        "title": "Week"
      },
...
    }
  }
}